vfio/type1: massage unmap iteration
mainline inclusion
from mainline-v5.12
commit 40ae9b807b89e60f86a40f33b931e6cde7eed2b7
category: feature

---------------------------------------------------------------------

Modify the iteration in vfio_dma_do_unmap so it does not depend on deletion
of each dma entry.  Add a variant of vfio_find_dma that returns the entry
with the lowest iova in the search range to initialize the iteration.  No
externally visible change, but this behavior is needed in the subsequent
update-vaddr patch.

Signed-off-by: Steve Sistare <steven.sistare@oracle.com>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>

Signed-off-by: luofei <luo.fei3@neolink.com>