sched/fair: Canceled the mandatory definition of offline tree
jingdong inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I80FQN

----------------------------------------------------

We have canceled the mandatory definition of offline tree.

for example:

A (root tg) -> B (online tg) -> C (offline tg) -> D (online tg) -> E (offline tg)

When we add the first task to D, the qos_idle_h_nr_running of D is 0,
but when the group se of to D is added to C, the corresponding qos_idle_h_nr_running becomes 1.
When returning upward, qos_idle_h_nr_running is both 1.

When a task is dequeued, D is dequeued as an online task,
but the group se corresponding to D is dequeued as an offline se.

Signed-off-by: zhangwei123171 <zhangwei123171@jd.com>
Reviewed-by: zhaoxiaoqiang11 <zhaoxiaoqiang11@jd.com>