ktask: add memory leak handling for ktask_works in ktask_init()
BUAA inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I85XYV

-----------------------------------------------------

The pointer ktask_works is allocated using kmalloc_array,
but is not freed in the error handling code.
This can lead to a memory leak if the initialization process fails.
To address this issue,
the allocated memory for the pointer ktask_works
should be freed in the respective error handling code.

Signed-off-by: Gui-Dong Han <hanguidong@buaa.edu.cn>