RDMA/core: fix kabi compatibility for SRQ resource tracing
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I86HQ2

--------------------------------------------------------------------------

In the patch related to SRQ resource tracing, members are
added to the structure, which causes Kapi incompatibility.

To solve this problem, perform the following operations:
First, make sure that the data structure of the new member
has little impact and the risk is controllable, then Use
the __GENKSYMS__ macro to shield the new structure members
and place the new members at the end of the data structure.

Fixes: d7b7ce53b3be ("RDMA/hns: Support SRQ restrack ops for hns driver")
Fixes: 788310f83ffc ("RDMA/core: Add support to dump SRQ resource in RAW format")
Fixes: 31cc28f1894f ("RDMA/core: Add dedicated SRQ resource tracker function")
Fixes: f54d69b7874e ("RDMA/nldev: Add QP numbers to SRQ information")
Fixes: 6102284160fd ("RDMA/nldev: Return SRQ information")
Fixes: 4e38b5a1267f ("RDMA/restrack: Add support to get resource tracking for SRQ")
Fixes: 0c32db26a6c4 ("RDMA/nldev: Return context information")

Signed-off-by: wenglianfa <wenglianfa@huawei.com>