radix tree: Don't return retry entries from lookup
mainline inclusion
from mainline-v4.20-rc7
commit eff3860bbfedbac6edac57fb0d7f3a60e860c1c3
category: bugfix
bugzilla:
CVE: NA

Reference: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=eff3860bbfedbac6edac57fb0d7f3a60e860c1c3
-------------------------------------------

Commit 66ee620f06f9 ("idr: Permit any valid kernel pointer to be stored")
changed the radix tree lookup so that it stops when reaching the bottom
of the tree.  However, the condition was added in the wrong place,
making it possible to return retry entries to the caller.  Reorder the
tests to check for the retry entry before checking whether we're at the
bottom of the tree.  The retry entry should never be found in the tree
root, so it's safe to defer the check until the end of the loop.

Add a regression test to the test-suite to be sure this doesn't come
back.

Fixes: 66ee620f06f9 ("idr: Permit any valid kernel pointer to be stored")
Reported-by: Greg Kurz <groug@kaod.org>
Signed-off-by: Matthew Wilcox <willy@infradead.org>

Conflict:
	tools/testing/radix-tree/Makefile

Signed-off-by: Ye Weihua <yeweihua4@huawei.com>