iommu/arm-smmu-v3: Integrate the function for obtain the device domain type in bypass mode
driver inclusion
category: cleanup
bugzilla: https://gitee.com/openeuler/kernel/issues/I5RP8T

-------------------------------------------------------------------

When the CONFIG_SMMU_BYPASS_DEV macro is enabled, the obtaining of domain
type is encapsulated as a function as a step in arm_smmu_device_domian_type
function. So that the code is clearer and easier to call later.

Signed-off-by: YunYi Yang <yangyunyi2@huawei.com>