crypto: hisilicon/qm - check function qp num before alg register
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I8IVG6
CVE: NA

----------------------------------------------------------------------

When the Kunpeng accelerator executes tasks such as encryption
and decryption have minimum requirements on the number of device
queues. If the number of queues does not meet the requirement,
the process initialization will fail. Therefore, the driver checks
the number of queues on the device before registering the algorithm.
If the number does not meet the requirements, the driver does not register
the algorithm to crypto subsystem, the device is still added to the
qm_list.

Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>