mm/sharepool: Implement mg_sp_alloc()
ascend inclusion
category: Feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8LNGH

---------------------------------------------

Alloc share memory for tasks in a share pool group. The tasks in the
same pool could access the memory with the same virtual address.

When the user calling mg_sp_alloc(), the physical memory is allocated
in do_mm_populate() for the first process and the pages are charged
into the memcg of that process. We should start at the current process
so as to charge the pages to the memcg of current process.

Signed-off-by: Wang Wensheng <wangwensheng4@huawei.com>