net: hns3: fix delete tc fail issue
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I8PDAE
CVE: NA

----------------------------------------------------------------------

When the tc is removed during reset, hns3 driver will return a errcode.
But kernel ignores this errcode, As a result,
the driver status is inconsistent with the kernel status.

This patch retains the deletion status when the deletion fails
and continues to delete after the reset to ensure that
the status of the driver is consistent with that of kernel.

Fixes: 46961be39152 ("net: hns3: add support for tc mqprio offload")
Signed-off-by: Jijie Shao <shaojijie@huawei.com>