mm/dynamic_pool: compatible with memory hwpoison
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8S9BY
CVE: NA

--------------------------------

When handle hwpoison on free pages in dpool, failed will eventually be
reported because they are not in buddy system or can not be dissolved.

But actually, it is safe because there are gatekeeper in allocation path
to make sure it can not be used and there are gatekeeper in demotion and
promotion path to make sure it can not pollute other pages.

Signed-off-by: Liu Shixin <liushixin2@huawei.com>