misc: add MISC_CG_RES_FD resource to control fd.
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8YC6O

----------------------------------------------------------------------

Filescgroup is used to control fd in cgroup v1 as cgroup controller,
which will not be merged to mainline. It is better to control fd with misc,
which provides the resource limiting and tracking mechanism for the scalar
resources and already be merged, so add MISC_CG_RES_FD resource to control
fd. Filescgroup will be abandoned in the future.

Signed-off-by: Chen Ridong <chenridong@huawei.com>