misc: add a type of resource to control fd.
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8YC6O

----------------------------------------------------------------------

Misc provides the resource limiting and tracking mechanism for the scalar
resources, it is more stable and concise to control fd than filescgroup,
which controls fd as independent cgroup, and filescgroup will be
abandoned in the future. Interfaces of misc are showed as below.

misc.capacity

A read-only flat-keyed file shown only in the root cgroup.  It shows
miscellaneous scalar resources available on the platform along with
their quantities::
	$ cat misc.capacity
	fd 1024

misc.current

A read-only flat-keyed file shown in the all cgroups.  It shows
the current usage of the resources in the cgroup and its children.

	$ cat misc.current
	fd 3

misc.max

A read-write flat-keyed file shown in the non root cgroups. Allowed
maximum usage of the resources in the cgroup and its children.

	$ cat misc.max
	fd max

	Limit can be set by::
	# echo fd 1 > misc.max

Signed-off-by: Chen Ridong <chenridong@huawei.com>