iomap: add IOMAP_DIO_INLINE_COMP
Rather than gate whether or not we need to punt a dio completion to a
workqueue on whether the IO is a write or not, add an explicit flag for
it. For now we treat them the same, reads always set the flags and async
writes do not.

No functional changes in this patch.

Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Conflicts:
	fs/iomap/direct-io.c
	[ 3a0be38cc84d("iomap: treat a write through cache the same
	  as FUA") is not applied.
	  a6d3d49587d1("iomap: switch __iomap_dio_rw to use iomap_iter")
	  is not applied. ]
Signed-off-by: Zhihao Cheng <chengzhihao1@huawei.com>