net: sfp: Synchronize some CMIS transceiver modules from ethtool
driver inclusion
category:feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9A3QT
CVE: NA

----------------------------------------------------------------------

Currently, the SFF-8024 Identifier Values that according to the standard
support for the Common Management Interface Specification (CMIS)
based on standard identifier values in the ethtool is more than in the
kernel. When the driver needs to use a newer Identifier Value, the kernel
interface does not support it. Therefore, we synchronize the CMIS mode
Identifier Values which supported by the ethtool to the kernel.

Signed-off-by: Hao Lan <lanhao@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>