crypto: hisilicon/qm - reset device before enabling it
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9CB6L
CVE: NA

----------------------------------------------------------------------

When the device is re-enabled, data that has not been cleared
may exist on the device. Therefore, reset the device to
the initial state before enabling the device. If an abnormal
interrupt event is reported when the device is enabled,
the device will not be reset.

Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>