crypto: hisilicon/qm - modify interrupt processing resource application
driver inclusion
category: cleanup
bugzilla: https://gitee.com/openeuler/kernel/issues/I9CB6L
CVE: NA

----------------------------------------------------------------------

The resources required by the interrupt processing like workqueue
are applied for in function hisi_qm_memory_init(). Resources are
requested regardless of whether interrupts are enabled. As a result,
the applied resources may not be used. To avoid waste of resources, the
resource application is moved to function qm_irqs_register(). When the
interrupt type is not supported, resource is not applied. In addition,
Interrupt registrations and interrupt resource application in the same
interface are easy to maintain.

Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>