crypto: hisilicon/hpre - mask cluster timeout error
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I9CB6L
CVE: NA

----------------------------------------------------------------------

The maximum timeout of the hpre cluster is 16 ms.
When a page fault occurs, the CPU core may process the
page fault event longer than 16 ms, especially when
the guest OS processes the page fault event reported
by the SMMU.

In the current test, there is a high probability that the
cluster times out. However, the cluster is waiting for the
completion of memory access, which is not an error. Therefore,
the device does not need to be reset. If an error occurs in the
cluster, the QM also reports the error. Therefore, the cluster
timeout error of hpre can be masked.

Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>